Move auth and dev route families into routes/auth/ and routes/dev/
domain/routes/ mixes one-file-per-entity route modules with two
hand-written families (the auth flow: auth_pages/auth_routes; the
dev-only routes: dev_auth/dev_components/dev_personas). Group each
family in a subpackage so the top level reads as "entity route files
plus single-file bespoke routes", per the updated README grammar.

Pure move + path updates (imports, contract-test patch target, seed
overrides, doc paths). `dev routes` before/after diff shows only the
handler module-path column changed — no URL or mount changes.
