+++ src/domain/routes/auth/__init__.py
@@ -1,0 +1,5 @@
+"""Hand-written auth route family — login/register/verify pages and the
+fastapi-users API routers. Not ``EntitySpec``-shaped, so nothing here
+registers on the entity registry; :mod:`src.main` imports these modules
+directly. See `../README.md` for the bespoke-routes rationale.
+"""
