+++ src/domain/routes/auth/auth_routes.py
