+++ src/domain/routes/auth/test_auth_routes.py
