+++ src/domain/templates/dev/components.html
@@ -2,7 +2,7 @@
    component with fixture data so they can be iterated in isolation.
 
    Route: GET /dev/components  (only mounted in ENVIRONMENT=development)
-   Template context: see src/domain/routes/dev_components.py
+   Template context: see src/domain/routes/dev/dev_components.py
 
    Gallery CSS is inlined here (not in framework.css / domain.css) since
 it's dev-only chrome that should never ship to production. #}
