+++ src/framework/http/README.md
@@ -17,7 +17,7 @@ Cross-cutting HTTP plumbing — error translation, response shaping, form parsin
 - The `_shared/form_meta.html::form_with_errors` macro wraps the `<form>` tag with the hx-* attributes the rerender path needs (`hx-target`, `hx-swap` — carrying a `show:window:top` scroll modifier so the re-render scrolls to the banner — and `hx-target-4xx`) plus the `{{ form_banner() }}` placement. There is no `hx-swap-4xx`: the `response-targets` extension overrides only the target on a 4xx, so the swap style comes from `hx-swap` (gotcha documented in `form_meta.html`). Templates `{% call form_with_errors(action=..., method=..., json_enc=...) %}` and supply the field inputs in the call block; the wrapper handles the rest. Replaces the 9-line wiring preamble that every opted-in form used to carry.
 - **Auth 401s take a separate path from the form-error rerender above.** The rerender path only catches exceptions raised *inside* a route body. An expired/absent session fails the `current_active_user` dependency, which raises `401` *before* the body runs — so no route-level handler is on the stack. These land at the global `unauthorized_exception_handler` in `src/main.py`, which routes by client shape to one login URL (`/auth/login?next=<path>`): htmx submit (`HX-Request: true`) → `204` + `HX-Redirect` (htmx navigates; no body for `hx-target-4xx` to swap); full-page browser (`Accept: text/html`) → `302`; JSON API → the `{"detail": ...}` body unchanged. The `HX-Request` branch is load-bearing: htmx sends `Accept: */*`, so without it an htmx submit falls through to the JSON arm and the raw `{"detail": ...}` gets swapped into the form. Pinned by `test_main.py::test_expired_session_*`.
 - `form_error_registry.py` — global `{ExceptionType: FormErrorMapping}` registry. Built-in registrations: `UserAlreadyExists` → 409 + `email` + "An account with this email already exists."; `InvalidPasswordException` → 422 + `password` + `e.reason`. Domain layers register their own typed exceptions at import time (next to the exception definition is the canonical home — see `_LoginBadCredentials` in `auth_pages.py`). XOR validation at registration time: a mapping must have exactly one of `field=`/`banner=True` and exactly one of `message=`/`message_from=` (callable for exceptions that carry their own detail).
-- `antibot.py` — anti-spam defense for the public auth forms (`/auth/register`, `/auth/forgot-password`). `enforce_antibot(request)` is called as the first line of a protected route body (before any side effect) and runs two layered checks: a **honeypot** (a hidden `contact_url` field — `HONEYPOT_FIELD` — that real users never fill; always checked, no key/network) and a gated **Cloudflare Turnstile** challenge (`verify_turnstile(...)` POSTs the widget token to Cloudflare's siteverify endpoint; runs only when `settings.CAPTCHA_ENABLED`, so tests/local-dev/contract clients pass without a key). Both failures raise `BotChallengeFailed`, a `BadRequestError` subclass registered in `form_error_registry` as a 400 banner — so HTMX submits get the "couldn't verify you're human" rerender and non-HTMX callers get a clean 400 (not a generic 500). `verify_turnstile` follows the `nppes.py` never-raise contract and *fails closed* (a captcha-service outage rejects rather than admits). The client half — the hidden input + widget markup — lives in `_shared/antibot.html` (`antibot_fields()` macro), fed by the `turnstile_site_key()` template global. Provider-specific bits (siteverify URL, `cf-turnstile-response` field, `success` payload) are isolated to this file + that macro, so an hCaptcha swap is local. Pinned by `test_antibot.py` plus route smokes in `src/domain/routes/test_auth_routes.py`.
+- `antibot.py` — anti-spam defense for the public auth forms (`/auth/register`, `/auth/forgot-password`). `enforce_antibot(request)` is called as the first line of a protected route body (before any side effect) and runs two layered checks: a **honeypot** (a hidden `contact_url` field — `HONEYPOT_FIELD` — that real users never fill; always checked, no key/network) and a gated **Cloudflare Turnstile** challenge (`verify_turnstile(...)` POSTs the widget token to Cloudflare's siteverify endpoint; runs only when `settings.CAPTCHA_ENABLED`, so tests/local-dev/contract clients pass without a key). Both failures raise `BotChallengeFailed`, a `BadRequestError` subclass registered in `form_error_registry` as a 400 banner — so HTMX submits get the "couldn't verify you're human" rerender and non-HTMX callers get a clean 400 (not a generic 500). `verify_turnstile` follows the `nppes.py` never-raise contract and *fails closed* (a captcha-service outage rejects rather than admits). The client half — the hidden input + widget markup — lives in `_shared/antibot.html` (`antibot_fields()` macro), fed by the `turnstile_site_key()` template global. Provider-specific bits (siteverify URL, `cf-turnstile-response` field, `success` payload) are isolated to this file + that macro, so an hCaptcha swap is local. Pinned by `test_antibot.py` plus route smokes in `src/domain/routes/auth/test_auth_routes.py`.
 - `middleware.py` — `StripEmptyQueryParamsMiddleware` (drops `?x=` empty pairs from form submissions so "filter not selected" behaves like "filter omitted"); `StaticNoCacheMiddleware` (dev-only — adds `Cache-Control: no-store` to `/static/*` so edits land without a hard refresh); `StaticLongCacheMiddleware` (production — emits `public, max-age=31536000, immutable` for `/static/...?v=<sha>` URLs and `public, max-age=300, must-revalidate` for unversioned ones; the `?v=` sentinel comes from the `static_version` Jinja global, sourced from `settings.APP_RELEASE`).
 - `static_files.py` — `MinifyingStaticFiles`, a `StaticFiles` subclass that runs `.css` responses through an in-process minifier and caches the bytes per path. Mounted at `/static/fw` and `/static/domain` in non-development environments (`src/main.py`). Conditional GETs honored via a weak ETag derived from the minified bytes so revalidating browsers get a 304.
