+++ tests/test_contract/infrastructure/servers/consumer.py
@@ -17,7 +17,7 @@ from src.auth_config import current_active_user, current_admin_user
 from src.domain import template_globals  # noqa: F401  # populates Jinja env globals
 from src.domain.logic.clinicians.schema import ClinicianCreate
 from src.domain.logic.programs.schema import ProgramCreate
-from src.domain.routes import auth_pages
+from src.domain.routes.auth import auth_pages
 from src.framework import APIResponse
 
 from ..utilities.mocks import MockAuthManager, create_mock_user
